+++ vm/src/hint_processor/builtin_hint_processor/dict_manager.rs
@@ -191,6 +191,17 @@ impl DictTracker {
         }
     }
 
+    //Returns a reference to the contained dictionary, losing the dictionary type in the process
+    pub fn get_dictionary_ref(&self) -> &HashMap<MaybeRelocatable, MaybeRelocatable> {
+        match &self.data {
+            Dictionary::SimpleDictionary(dict) => dict,
+            Dictionary::DefaultDictionary {
+                dict,
+                default_value: _,
+            } => dict,
+        }
+    }
+
     pub fn get_value(&mut self, key: &MaybeRelocatable) -> Result<&MaybeRelocatable, HintError> {
         self.data
             .get(key)