+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 #### Upcoming Changes
 
+#### [0.9.3] - 2024-06-07
+
 * feat: Add `EXCESS_BALANCE` hint [#1777](https://github.com/lambdaclass/cairo-vm/pull/1777)
 
-#### [0.9.2] - 2024-01-3
+#### [0.9.2] - 2024-01-03
 
 * Change ec_op_impl() to use ProjectivePoint [#1534](https://github.com/lambdaclass/cairo-vm/pull/1534)